server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-cloud-storage==2.7.0
werkzeug==2.0.3
pymongo==4.6.2
python-dotenv==1.0.1 